new migration

+++ templates/templateapp/app/myself/myself_view.go
@@ -10,7 +10,7 @@ import (
 )
 
 type Json struct {
-	ID        uint      ` + "`" + `json:"id"` + "`" + `
+	ID        uint64    ` + "`" + `json:"id"` + "`" + `
 	Name      string    ` + "`" + `json:"name"` + "`" + `
 	Email     string    ` + "`" + `json:"email"` + "`" + `
 	Locale    string    ` + "`" + `json:"locale"` + "`" + `
