+++ templates/templateapp/app/usertemplate/user_view.go
@@ -21,7 +21,7 @@ type SuccessfullySavedJson struct {
 }
 
 type Json struct {
-	ID        uint       ` + "`" + `json:"id"` + "`" + `
+	ID        uint64     ` + "`" + `json:"id"` + "`" + `
 	Name      string     ` + "`" + `json:"name"` + "`" + `
 	Email     string     ` + "`" + `json:"email"` + "`" + `
 	Admin     bool       ` + "`" + `json:"admin"` + "`" + `
